updated design of webpage
Added links inside the <head> in order to use fonts and icons. Created a grid with hover text effects including a dropdown menu that lead to different webpages. Customized CSS to my liking while also fixing bugs and also included a neon fade back ground to the grid. A new folder was also created for the webpage background or any images that will be added in the future.