changed column topic & installed bootstrap
Used bootstrap to install an automatic playing carousel to webpage to do that I used a main slider container div that also included the fade effect to the images.
A wrapper div for each slide, containing three images.
Two buttons for navigating between the slides.
I then adjusted the positioning of H1, row, carousel, and d-block to fit the page perfectly.